Added Authorization for controllers

+++ FinanceManager/Controllers/v1/TransactionController.cs
@@ -4,7 +4,10 @@ using FinanceManager.Extensions;
 using FinanceManager.Requests.v1;
 using FinanceManager.Responses.v1;
 using FinanceManager.Services;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration.UserSecrets;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -13,6 +16,7 @@ namespace FinanceManager.Controllers.v1
     [Route("api/v{version:apiVersion}/transactions/")]
     [ApiController]
     [ApiVersion("1.0")]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public class TransactionController : ControllerBase
     {
         private readonly ITransactionService _transactionService;
@@ -36,6 +40,12 @@ namespace FinanceManager.Controllers.v1
         [HttpGet("{transactionId:int}", Name = "GetTransaction")]
         public async Task<IActionResult> Get(int transactionId)
         {
+            var userId = HttpContext.GetUserId();
+            var isValidUser = await _transactionService.IsTransactionOwnerAsync(transactionId, userId);
+
+            if (!isValidUser)
+                return BadRequest(new { error = "User is not transaction owner." });
+
             var transaction = await _transactionService.GetTransactionByIdAsync(transactionId);
 
             if (transaction == null)
@@ -53,6 +63,7 @@ namespace FinanceManager.Controllers.v1
                 return BadRequest(ModelState.GetErrorMessages());
 
             var transaction = _mapper.Map<Transaction>(transactionRequest);
+            transaction.UserId = HttpContext.GetUserId();
 
             var createdTransaction = await _transactionService.CreateTransactionAsync(transaction);
 
@@ -68,11 +79,18 @@ namespace FinanceManager.Controllers.v1
         [HttpPut("{transactionId:int}")]
         public async Task<IActionResult> Update([FromRoute] int transactionId, [FromBody] UpdateTransactionRequest transactionRequest)
         {
+            var userId = HttpContext.GetUserId();
+            var isValidUser = await _transactionService.IsTransactionOwnerAsync(transactionId, userId);
+
+            if (!isValidUser)
+                return BadRequest(new { error = "User is not transaction owner." });
+
             if (!ModelState.IsValid)
                 return BadRequest(ModelState.GetErrorMessages());
 
             var transaction = _mapper.Map<Transaction>(transactionRequest);
             transaction.Id = transactionId;
+            transaction.UserId = userId;
 
             var updated = await _transactionService.UpdateTransactionAsync(transaction);
 
@@ -87,6 +105,12 @@ namespace FinanceManager.Controllers.v1
         [HttpDelete("{transactionId:int}")]
         public async Task<IActionResult> Delete([FromRoute] int transactionId)
         {
+            var userId = HttpContext.GetUserId();
+            var isValidUser = await _transactionService.IsTransactionOwnerAsync(transactionId, userId);
+
+            if (!isValidUser)
+                return BadRequest(new { error = "User is not transaction owner." });
+
             var deleted = await _transactionService.DeleteTransactionAsync(transactionId);
 
             if (!deleted)
